Simplify header format. Parse header into map.
The header format was dumbed down a bit to be required, just so
I could get this working. And it mostly does.

After a lot of false starts, take-header will produce the header lines,
and then header-lines-as-map will transform it into a map (duh).

It took some trial and error since it's been a while since I've done
Clojure. The fact that I'm using Emacs is sort of a wash; the REPL
integration is fantastic, but I'm still pretty bumbling.